moved LanguageUtils to Utils. + split getDeviceLanguage into Code (for string) and Locale for reuse as suggested in code review by @maxme
